signout/registration
currently-signout button closes entire program, we will need to add a flag section in the registeredusers.csv file to indicate if the user is online.

registration - adjusted in the server file just the way the registeredusers.csv file was opened because I was getting a permissions error when running it originally. I commented out the original code that was there.

Also for registration- there are some validation issues for email and username so far. It might just be a sequence issue with the GUI. I will play with it some more.